pyproject.toml
[project]
name = "yt-dlp-dl"
version = "0.1.0"
description = "Use yt-dlp to download videos in global env by alias in Windows PowerShell"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "requests>=2.32.5",
    "uv>=0.9.28",
    "yt-dlp>=2025.12.8",
    "python-dotenv>=1.0.0",
]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[[tool.uv.index]]
name = "baidu"
url = "https://mirror.baidu.com/pypi/simple/"

[[tool.uv.index]]
name = "ustc"
url = "https://pypi.mirrors.ustc.edu.cn/simple/"
